requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
atproto==0.0.62
attrs==25.4.0
black==25.9.0
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==45.0.7
dnspython==2.8.0
Flask==3.1.2
Flask-Caching==2.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
libipld==3.1.1
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
narwhals==2.8.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.0
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==13.1
Werkzeug==3.1.3
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
atproto==0.0.62
attrs==25.4.0
black==25.9.0
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==45.0.7
dnspython==2.8.0
Flask==3.1.2
Flask-Caching==2.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
libipld==3.1.1
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
narwhals==2.8.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.0
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==13.1
Werkzeug==3.1.3
